Update optimizer_botorch.py
If there are categorical variables, the new columns must be of dtype float64, otherwise the sampling step fails because the numpy array holding the dataframe is of type object, which is incompatible with the scaling steps in the sampling.

+++ edbo/plus/optimizer_botorch.py
@@ -58,7 +58,7 @@ class EDBOplus:
                   f" using One-Hot-Encoding: {ohe_columns}")
         # Encode OHE.
         df_sampling = pd.get_dummies(df, prefix=ohe_columns,
-                                     columns=ohe_columns, drop_first=True)
+                                     columns=ohe_columns, drop_first=True, dtype=np.float64)
 
         # Order df according to initial sampling method (random samples).
         idaes = None
